Project directory cleanup (#1514)
* Removed erroneous duplicate project directory
* Corrected project name for consistency across test projects
