+++ test/Dapr.Actors.Analyzers.Test/VerifyCodeFix.cs
@@ -1,13 +1,26 @@
-﻿using Microsoft.CodeAnalysis.CodeActions;
+﻿// ------------------------------------------------------------------------
+//  Copyright 2025 The Dapr Authors
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//      http://www.apache.org/licenses/LICENSE-2.0
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//  ------------------------------------------------------------------------
+
+using Microsoft.CodeAnalysis.CodeActions;
 using Microsoft.CodeAnalysis.CodeFixes;
 
-namespace Dapr.Actors.Analyzers.Test;
+namespace Dapr.Actors.Analyzers.Tests;
 
-internal class VerifyCodeFix
+internal static class VerifyCodeFix
 {
     public static async Task RunTest<T>(string code, string expectedChangedCode) where T : CodeFixProvider, new()
     {
-        var (diagnostics, document, workspace) = await Utilities.GetDiagnosticsAdvanced(code);
+        var (diagnostics, document, workspace) = await TestUtilities.GetDiagnosticsAdvanced(code);
 
         Assert.Single(diagnostics);
 
@@ -41,16 +54,17 @@ internal class VerifyCodeFix
         var updatedDocument = workspace.CurrentSolution.GetDocument(document.Id) ?? throw new Exception("Updated document is null");
         var newCode = (await updatedDocument.GetTextAsync()).ToString();
 
+        var normalizedExpectedCode = NormalizeWhitespace(expectedChangedCode);
+        var normalizedNewCode = NormalizeWhitespace(newCode);
+
+        Assert.Equal(normalizedExpectedCode, normalizedNewCode);
+        return;
+
         // Normalize whitespace
         string NormalizeWhitespace(string input)
         {
             var separator = new[] { ' ', '\r', '\n' };
             return string.Join(" ", input.Split(separator, StringSplitOptions.RemoveEmptyEntries));
         }
-
-        var normalizedExpectedCode = NormalizeWhitespace(expectedChangedCode);
-        var normalizedNewCode = NormalizeWhitespace(newCode);
-
-        Assert.Equal(normalizedExpectedCode, normalizedNewCode);
     }
 }
